feat: split filters
