+++ packages/frontend2/src/app/(new)/(other)/scaling/data-availability/_components/scaling-da-filters.tsx
@@ -1,21 +1,16 @@
-import { type StageConfig } from '@l2beat/config'
-import { compact, uniq } from 'lodash'
+import { uniq } from 'lodash'
 import React from 'react'
-import { Checkbox } from '~/app/_components/checkbox'
-import { OverflowWrapper } from '~/app/_components/overflow-wrapper'
 import { TableFilter } from '~/app/_components/table/filters/table-filter'
 import { type ScalingDataAvailabilityEntry } from '~/server/features/scaling/types'
+import {
+  BaseScalingFilters,
+  type BaseScalingFiltersState,
+} from '../../../_components/base-scaling-filters'
 
 type ScalingFiltersEntry = ScalingDataAvailabilityEntry
-
-export interface ScalingDaFiltersState {
-  rollupsOnly: boolean | undefined
-  category: string | undefined
-  stack: string | undefined
-  stage: string | undefined
-  purpose: string | undefined
+export type ScalingDaFiltersState = {
   daLayer: string | undefined
-}
+} & BaseScalingFiltersState
 
 interface Props {
   items: ScalingFiltersEntry[]
@@ -23,44 +18,7 @@ interface Props {
   setState: React.Dispatch<React.SetStateAction<ScalingDaFiltersState>>
 }
 
-// TODO: Merge/split into components since scaling-filter are closed in terms of logic
 export function ScalingDaFilters({ items, state, setState }: Props) {
-  const typeOptions = uniq(items.map((item) => item.category))
-    .sort()
-    .map((value) => ({
-      label: value,
-      value,
-    }))
-
-  const stackOptions = uniq(items.map((item) => item.provider))
-    .sort()
-    .map((value) => ({
-      label: value ?? 'No stack',
-      value,
-    }))
-
-  const stageOptions = uniq(
-    compact(
-      items.map((item) => {
-        if ('stage' in item) {
-          return item.stage?.stage
-        }
-      }),
-    ),
-  )
-    .sort()
-    .map((value) => ({
-      label: stageLabel(value),
-      value,
-    }))
-
-  const purposeOptions = uniq(items.flatMap((item) => item.purposes))
-    .sort()
-    .map((value) => ({
-      label: value,
-      value,
-    }))
-
   const daLayerOptions = uniq(
     items.flatMap((item) => item.dataAvailability.layer.value),
   )
@@ -70,72 +28,23 @@ export function ScalingDaFilters({ items, state, setState }: Props) {
       value,
     }))
 
-  const isRollupInItems = items.some((item) => item.category.includes('Rollup'))
+  const daLayerFilter = (
+    <TableFilter
+      title="DA Layer"
+      options={daLayerOptions}
+      value={state.daLayer}
+      onValueChange={(value) =>
+        setState((prev) => ({ ...prev, daLayer: value }))
+      }
+    />
+  )
 
   return (
-    <OverflowWrapper>
-      <div className="flex space-x-2">
-        <Checkbox
-          id="rollups-only"
-          onCheckedChange={(checked) =>
-            setState((prev) => ({ ...prev, rollupsOnly: !!checked }))
-          }
-          disabled={!isRollupInItems}
-        >
-          Rollups only
-        </Checkbox>
-        <TableFilter
-          title="Type"
-          options={typeOptions}
-          value={state.category}
-          onValueChange={(value) =>
-            setState((prev) => ({ ...prev, category: value }))
-          }
-        />
-        <TableFilter
-          title="Stack"
-          options={stackOptions}
-          value={state.stack}
-          onValueChange={(value) =>
-            setState((prev) => ({ ...prev, stack: value }))
-          }
-        />
-        <TableFilter
-          title="Stage"
-          options={stageOptions}
-          value={state.stage}
-          onValueChange={(value) =>
-            setState((prev) => ({ ...prev, stage: value }))
-          }
-        />
-        <TableFilter
-          title="Purpose"
-          options={purposeOptions}
-          value={state.purpose}
-          onValueChange={(value) =>
-            setState((prev) => ({ ...prev, purpose: value }))
-          }
-        />
-        <TableFilter
-          title="DA Layer"
-          options={daLayerOptions}
-          value={state.daLayer}
-          onValueChange={(value) =>
-            setState((prev) => ({ ...prev, daLayer: value }))
-          }
-        />
-      </div>
-    </OverflowWrapper>
+    <BaseScalingFilters
+      setState={setState}
+      state={state}
+      items={items}
+      additionalFilters={daLayerFilter}
+    />
   )
-}
-
-function stageLabel(stage: StageConfig['stage']) {
-  switch (stage) {
-    case 'NotApplicable':
-      return 'Not applicable'
-    case 'UnderReview':
-      return 'Under review'
-    default:
-      return stage
-  }
 }
